02/07

+++ dev/Domain/CategoryRepository.java
@@ -25,34 +25,50 @@ public class CategoryRepository  implements IRepository<Category>  {
     }
 
     @Override
-    public void add(Category category) throws Exception{
-        if(categories.containsKey(category.getCategoryName())) {
-            throw new Exception("Category " + category.getCategoryName() + " already exist.");
+    public void add(Category category) throws Exception {
+        if (categories.containsKey(category.getCategoryName())) {
+            throw new Exception("Category " + category.getCategoryName() + " already exists.");
+        } else {
+            Category categoryFromDB = categoryDAO.get(category.getCategoryName());
+            if (categoryFromDB != null) {
+                categories.put(category.getCategoryName(), categoryFromDB);
+            } else {
+                categoryDAO.add(category);
+                categories.put(category.getCategoryName(), category);
+            }
         }
-        categories.put(category.getCategoryName(), category);
-        categoryDAO.add(category);
     }
 
     @Override
-    public void remove(String name)  throws Exception{
-        catIsExist(name);
-        categories.remove(name);
-        categoryDAO.remove(name);
+    public void remove(String name) throws Exception {
+        if (!categories.containsKey(name)) {
+            Category categoryFromDB = categoryDAO.get(name);
+            if (categoryFromDB == null) {
+                throw new Exception("Category " + name + " doesn't exist.");
+            } else {
+                categoryDAO.remove(name);
+            }
+        } else {
+            categories.remove(name);
+            categoryDAO.remove(name);
+        }
     }
 
     @Override
-    public Category get(String name) throws Exception{
+    public Category get(String name) throws Exception {
         Category category = categories.get(name);
-        if (category == null) {
+        if (category != null) {
+            return category;
+        } else {
             category = categoryDAO.get(name);
             if (category != null) {
                 categories.put(name, category);
+                return category;
+            } else {
+                throw new Exception("Category " + name + " doesn't exist.");
             }
         }
-        catIsExist(name);
-        return category;
     }
-
     public void showCategories() throws SQLException{
         categoryDAO.showCategories();
     }
@@ -73,7 +89,9 @@ public class CategoryRepository  implements IRepository<Category>  {
         }
         return true;
     }
-
+    public void close() throws SQLException {
+        categoryDAO.close();
+    }
 
 
 }
